refactor(html): tree construction "template"
